More compiler warnings

+++ Base/test/TestVector.cpp
@@ -146,8 +146,8 @@ TEST(VectorBase, DotProduct) {
   EXPECT_EQ(v5i.Dot(v5u), static_cast<int>(10));
   EXPECT_EQ(v5u.Dot(v5i), static_cast<unsigned>(10));
 
-  EXPECT_EQ(v5i * v5u, 10);
-  EXPECT_EQ(v5u * v5i, 10);
+  EXPECT_EQ(v5i * v5u, static_cast<int>(10));
+  EXPECT_EQ(v5u * v5i, static_cast<unsigned>(10));
 }
 
 TEST(VectorBase, Length) {
